Add function to check gameboard diagonals for winner
Check whether any diagonal on the board contains the same element, and
ensure it is not null (otherwise a diagonal of nulls would count as a
win).

While there are ways to check diagonals via algorithm/looping, a brute
force check is simpler, and since the only grid the function handles is
going to be a 3x3 grid, scalability is not a concern.