Move gameboard object creation into createGameController
Change gameboard from IIFE to factory function and call it exclusively
within createGameController, returning the gameboard's relevant methods
through createGameController.

gameboard and gameController objects have no meaningful data to store
after a round is completed. To start a new round, they must be reset or
created from scratch.

It may be more efficient to re-use and reset them, but given how small
the objects (and the size of this project) are, the simplicity of not
worrying about reset methods is worth the tradeoff.