Refactor gameboard to use 1D array
Use a 1D array (e.g. board[cells]) instead of a 2D array (e.g.
board[row][col]) for the board.

This greatly simplifies code interacting with the board: it eliminates
several nested loops, and reduces the parameters for interacting with a
board cell from (row, col) to just (cell).

In line with this change, algorithmic checks for wins (rowHasWinner,
colHasWinner, and diagonalHasWinner) have been removed and replaced with
brute-force checks for all win conditions. This also greatly reduces the
amount code involved with win checks and is much simpler to understand.

Removing and replacing the win check functions could warrant a commit on
its own, but is included here because:

  1. The game breaks without win check functions
  2. The old win check functions are not compatible with this refactor

Since this is a required change for the 1D array refactor to work, I
have decided to include it here.

Change inspired by mujahidshaik724's tic-tac-toe project, which can be
found below:

https://github.com/mujahidshaik724/TIC-TAC-TOE