Refactor gameboard to store player ids instead of icons
Store player.id in gameboard cells instead of player.icon. Practically
speaking, the impact this has on the project is basically none. However,
this change separates the values gameboard logic uses from the values
displayed to the user.

The gameboard logic and displayController logic no longer need to handle
the same value, something which, in a larger project, may have been more
problematic.